java-src/main/java/com/zhangliang/leetcode/LongestPalindromicSubsequence.java
package com.zhangliang.leetcode;
/*
Given a string s, find the longest palindromic subsequence's 
length in s. You may assume that the maximum length of s is 1000.

Example 1:
Input:

"bbbab"
Output:
4
One possible longest palindromic subsequence is "bbbb".

Example 2:
Input:

"cbbd"
Output:
2
One possible longest palindromic subsequence is "bb".
*/

public class LongestPalindromicSubsequence {
    public int longestPalindromeSubseq(String s) {
        int[][] dp = new int[s.length()][s.length()];
        for (int len = 1; len <= s.length(); len++) {
            for (int i = 0; i + len - 1 < s.length(); i++) {
                int j = i + len - 1;
                if (s.charAt(i) == s.charAt(j)) {
                    if (len > 2) {
                        dp[i][j] = dp[i + 1][j - 1] + 2;
                    } else if (len < 2) {
                        dp[i][j] = 1;
                    } else {
                        dp[i][j] = 2;
                    }
                } else {
                    dp[i][j] = Math.max(dp[i + 1][j], dp[i][j - 1]);
                }
            }
        }
        return dp[0][s.length() - 1];
    }
}
